refactor: split math generation into individual files
